Use uint32 for chainID in chain registration
These functions were taking a mix of uint16 and uint32 for the chain id,
which is just a bug waiting to happen.  Use uint32 for all functions
because that's what the protobuf uses.

The chain id is still encoded as a uint16 in the various payloads but
it's trivial to cast a uint16 to a uint32 whereas we would need to
insert a bounds check every time we go the other way.

commit-id:42e6fe36

+++ wormhole_chain/x/tokenbridge/keeper/chain_registration.go
@@ -18,14 +18,12 @@ func (k Keeper) SetChainRegistration(ctx sdk.Context, chainRegistration types.Ch
 // GetChainRegistration returns a chainRegistration from its index
 func (k Keeper) GetChainRegistration(
 	ctx sdk.Context,
-	chainID uint16,
+	chainID uint32,
 
 ) (val types.ChainRegistration, found bool) {
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefix(types.ChainRegistrationKeyPrefix))
 
-	b := store.Get(types.ChainRegistrationKey(
-		uint32(chainID),
-	))
+	b := store.Get(types.ChainRegistrationKey(chainID))
 	if b == nil {
 		return val, false
 	}
